refactor(nav-pilot): extract internal/domain (P2)
- Move Config, ResolvedConfig, CLIOverrides, ModelChoice, valid* slices,
  validateModelValue/OptionalModel, containsStr, modelValuePattern to
  internal/domain
- Move InstallScope + ScopeRepo/ScopeUser to internal/domain
- Move StateFile, InstalledFile, FileStatus* consts to internal/domain
- Move color/output helpers (Bold/Dim/Green/Red/Yellow) to internal/domain
- Add aliases.go in package main (type aliases + var aliases) so zero
  other files need editing
- Delete output.go and scope.go (content fully in domain now)
- Trim config.go and state.go of moved declarations
