+++ cli/nav-pilot/internal/domain/domain.go
@@ -1,11 +1,113 @@
-package main
+package domain
 
 import (
+	"errors"
 	"fmt"
 	"os"
 	"path/filepath"
+	"regexp"
 	"strings"
 )
+
+// Config holds user-specific nav-pilot configuration read from ~/.nav-pilot/config.toml.
+// Pointer types are used for optional fields to distinguish "unset" from zero-value,
+// enabling correct per-field precedence in resolve().
+type Config struct {
+	Version         int     `toml:"version"`
+	Client          *string `toml:"client"`
+	Model           *string `toml:"model"`
+	Mode            *string `toml:"mode"`
+	ReasoningEffort *string `toml:"reasoning_effort"`
+	ContextTier     *string `toml:"context_tier"`
+	AllowAllTools   *bool   `toml:"allow_all_tools"`
+	AskUser         *bool   `toml:"ask_user"`
+	LogLevel        *string `toml:"log_level"`
+	OtelLogLevel    *string `toml:"otel_log_level"`
+}
+
+// ResolvedConfig holds the final configuration after applying precedence:
+// CLI flag > file value > built-in default.
+type ResolvedConfig struct {
+	Client          string
+	Model           string // empty = use agent default
+	Mode            string
+	ReasoningEffort string // empty = unset
+	ContextTier     string // empty = unset
+	AllowAllTools   bool
+	AskUser         bool
+	LogLevel        string // empty = unset
+	OtelLogLevel    string // always set; defaults to "none"
+}
+
+// CLIOverrides holds optional CLI flag values. Empty string means "not provided via CLI".
+type CLIOverrides struct {
+	Client          string
+	Model           string
+	Mode            string
+	ReasoningEffort string
+	ContextTier     string
+	AllowAllTools   *bool
+	AskUser         *bool
+	LogLevel        string
+	OtelLogLevel    string
+}
+
+var (
+	ValidModes           = []string{"default", "plan", "autopilot"}
+	ValidReasoningEffort = []string{"none", "low", "medium", "high", "xhigh", "max"}
+	ValidContextTiers    = []string{"default", "long_context"}
+	ValidLogLevels       = []string{"none", "error", "warning", "info", "debug", "all", "default"}
+	ValidOtelLogLevels   = []string{"none", "error", "warning", "warn", "info", "debug", "verbose", "all"}
+)
+
+// ModelChoice pairs a model id (the --model value) with a human-readable label.
+// The concrete lists (knownCopilotModels, knownOpenCodeModels) live in clients.go.
+type ModelChoice struct {
+	ID    string
+	Label string
+}
+
+// ModelValuePattern restricts model identifiers to a sane character set that
+// covers Copilot ids (e.g. "claude-opus-4.8", "gpt-5.5") and opencode
+// provider/model ids (e.g. "anthropic/claude-3-5-sonnet").
+var ModelValuePattern = regexp.MustCompile(`^[A-Za-z0-9][A-Za-z0-9._/-]*$`)
+
+// ValidateModelValue applies strong format validation to a model identifier.
+// The model catalog is dynamic (Copilot validates server-side), so this checks
+// shape rather than membership: non-empty, no surrounding/inner whitespace, and
+// a restricted character set. This rejects typos and garbage while remaining
+// correct as the model catalog evolves.
+func ValidateModelValue(model string) error {
+	if strings.TrimSpace(model) != model {
+		return fmt.Errorf("model %q must not have leading or trailing whitespace", model)
+	}
+	if model == "" {
+		return errors.New("model must not be empty (omit the key to use the agent default)")
+	}
+	if !ModelValuePattern.MatchString(model) {
+		return fmt.Errorf("model %q is not a valid identifier (allowed characters: letters, digits, '.', '_', '-', '/')", model)
+	}
+	return nil
+}
+
+// ValidateOptionalModel is a huh form validator: it accepts a blank value
+// (meaning "unset / agent default") and otherwise applies ValidateModelValue.
+func ValidateOptionalModel(s string) error {
+	s = strings.TrimSpace(s)
+	if s == "" {
+		return nil
+	}
+	return ValidateModelValue(s)
+}
+
+func ContainsStr(list []string, s string) bool {
+	for _, v := range list {
+		if v == s {
+			return true
+		}
+	}
+	return false
+}
 
 // InstallScope encapsulates the differences between repo-level and user-level installs.
 type InstallScope struct {
@@ -164,3 +266,49 @@ func (s *InstallScope) Label() string {
 func (s *InstallScope) IsUser() bool {
 	return s.Name == "user"
 }
+
+// StateFile tracks what was installed, for safe updates and uninstall.
+type StateFile struct {
+	Collection  string          `json:"collection"`
+	Version     string          `json:"version"`
+	Scope       string          `json:"scope,omitempty"` // "repo" or "user"; empty means "repo" (backwards compat)
+	SourceSHA   string          `json:"source_sha"`
+	InstalledAt string          `json:"installed_at"`
+	Files       []InstalledFile `json:"files"`
+}
+
+// InstalledFile records a single installed file with its content hash.
+type InstalledFile struct {
+	Path   string `json:"path"`
+	Hash   string `json:"hash"`
+	Status string `json:"status,omitempty"` // "" = active, FileStatusIgnored = intentionally excluded, FileStatusConflict = exists with local modifications
+}
+
+// FileStatusIgnored marks a file as intentionally excluded by the user.
+// Sync and status skip files with this status.
+const FileStatusIgnored = "ignored"
+
+// FileStatusConflict marks a file that existed with local modifications at install time.
+// The user declined to overwrite it, so sync should not touch it until resolved.
+const FileStatusConflict = "conflict"
+
+var UseColor = true
+
+func init() {
+	if os.Getenv("NO_COLOR") != "" {
+		UseColor = false
+	}
+}
+
+func Color(code, msg string) string {
+	if !UseColor {
+		return msg
+	}
+	return fmt.Sprintf("\033[%sm%s\033[0m", code, msg)
+}
+
+func Red(msg string) string    { return Color("31", msg) }
+func Green(msg string) string  { return Color("32", msg) }
+func Yellow(msg string) string { return Color("33", msg) }
+func Dim(msg string) string    { return Color("2", msg) }
+func Bold(msg string) string   { return Color("1", msg) }
